+++ conkit/applications/ccmpred.py
@@ -35,7 +35,7 @@ Command line object for CCMpred contact prediction application
 
 __author__ = "Felix Simkovic"
 __date__ = "04 Aug 2016"
-__version__ = "0.13.3"
+__version__ = "0.14.1"
 
 from conkit.applications._compat import _Argument, _Option, _Switch, AbstractCommandline